+++ GameTemplate/k2EngineLow/physics/DebugWireframe.cpp
@@ -154,7 +154,7 @@ void DebugWireframe::RenderContextUpdate(RenderContext& rc)
 	//ディスクリプタヒープに設定。
 	rc.SetDescriptorHeap(m_descriptorHeap);
 	//ドローコール。
-	rc.DrawIndexed(m_vertexList.size());
+	rc.DrawIndexed(static_cast<UINT>(m_vertexList.size()));
 }
 
 //1フレーム内にdrawLineは線の数だけ行う
